chore: renaming and add typing
