+++ bc_obps/registration/api/users.py
@@ -6,9 +6,9 @@ from registration.api.api_base import router
 ##### GET #####
 
 
-@router.get("/user", response=UserOut, url_name="get_user_by_guid")
+@router.get("/user", response=UserOut, url_name="get_by_guid")
 @authorize(AppRole.get_all_authorized_app_roles(), UserOperator.get_all_industry_user_operator_roles(), False)
-def get_user_by_guid(request):
+def get_by_guid(request):
     user: User = request.current_user
     return user
