+++ bc_obps/registration/tests/endpoints/test_user_app_role_endpoints.py
@@ -2,5 +2,5 @@ from registration.tests.utils.helpers import CommonTestSetup
 
 
 class TestUserAppRoleEndpoint(CommonTestSetup):
-    def test_get_user_app_role(self):
+    def test_get_app_role(self):
         pass
