+++ bc_obps/registration/tests/endpoints/test_user_endpoints.py
@@ -5,13 +5,13 @@ from registration.utils import custom_reverse_lazy
 class TestUsersEndpoint(CommonTestSetup):
     def test_unauthorized_users_cannot_get(self):
         # /user
-        response = TestUtils.mock_get_with_auth_role(self, "cas_pending", custom_reverse_lazy('get_user_by_guid'))
+        response = TestUtils.mock_get_with_auth_role(self, "cas_pending", custom_reverse_lazy('get_by_guid'))
         assert response.status_code == 401
 
     # GET USER
-    def test_get_user_by_guid(self):
+    def test_get_by_guid(self):
         # Act
-        response = TestUtils.mock_get_with_auth_role(self, 'industry_user', custom_reverse_lazy('get_user_by_guid'))
+        response = TestUtils.mock_get_with_auth_role(self, 'industry_user', custom_reverse_lazy('get_by_guid'))
         content = response.json()
         # Assert
         assert response.status_code == 200
